requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0
Pillow==12.1.1
requests==2.33.0
PyJWT==2.12.1
sqlparse==0.5.5
asgiref==3.11.1
uritemplate==4.2.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
PyYAML==6.0.3
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
urllib3==2.6.3
tzdata==2025.3
